requirements.txt
openpyxl==3.1.2
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.0.1